chunk SPI DMA xfers to support larger buffers

+++ src/drivers/display.c
@@ -104,7 +104,7 @@ void display_pixel_stream_end(void) {
   end_tx();
 }
 
-void display_pixel_stream_write(const uint16_t *pixels, uint16_t count) {
+void display_pixel_stream_write(const uint16_t *pixels, uint32_t count) {
   // SPI must operate in 8-bit mode although pixels must be 16bits.
   spi_tx(io.spi, (uint8_t *)pixels, count * 2);
 }
