+++ src/drivers/display.h
@@ -17,5 +17,5 @@ void display_set_window(uint16_t x0, uint16_t y0, uint16_t x1, uint16_t y1);
 
 void display_pixel_stream_begin(void);
 // NOTE: pixel_stream is assumed to be in big endian
-void display_pixel_stream_write(const uint16_t *pixels, uint16_t count);
+void display_pixel_stream_write(const uint16_t *pixels, uint32_t count);
 void display_pixel_stream_end(void);
